Create generic struture for the build and test
In the Makefile was created a variable NAMES that defines the
target files for build and running in SDE emulator.

It is possible to increase the number of runnings of each test in
Makefile changing the var REPEAT which perf should run the N times
each test and collect the average.

Signed-off-by: Gustavo Walbon <gustavowalbon@gmail.com>